requirements.txt
asgiref==3.8.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
mypy==1.8.0
mypy-extensions==1.0.0
packaging==24.1
pillow==10.4.0
progressbar==2.5
psycopg2-binary==2.9.9
PyJWT==2.9.0
requests==2.31.0
sqlparse==0.5.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
PyPDF2~=3.0.1